refactor(client): :art: small naming fixes
