+++ packages/client/src/components/FireboltForm/InputHolder/hook/helpers.ts
@@ -1,3 +1,5 @@
+import { validateFBTField } from "@iq-firebolt/validators/src"
+import { IStepConfigField } from "@iq-firebolt/client-core/lib"
 import evaluate from "simple-evaluate"
 
 export function getConditionalProps({ propsConditional, formPayload }) {
@@ -14,4 +16,18 @@ export function getConditionalProps({ propsConditional, formPayload }) {
     },
     {}
   )
+}
+
+interface IvalidateField {
+  value: string
+  field: IStepConfigField
+  formPayload: { [fieldSlug: string]: any }
+}
+export function validateField({ value, field, formPayload }: IvalidateField) {
+  return validateFBTField({
+    value,
+    field,
+    formPayload,
+    context: "client",
+  })
 }
