+++ packages/client/src/components/FireboltForm/InputHolder/index.tsx
@@ -2,13 +2,13 @@ import React from "react"
 import useInputHolder, { IUseInputHolder } from "./hook"
 
 const InputHolder = (props: IUseInputHolder) => {
-  const { computedClasses, slug, nfieldProps } = useInputHolder(props)
+  const { computedClasses, slug, fieldProps } = useInputHolder(props)
 
   const { FieldComponent } = props
 
   return (
     <div className={computedClasses} data-fieldslug={slug}>
-      <FieldComponent {...nfieldProps} />
+      <FieldComponent {...fieldProps} />
     </div>
   )
 }
