+++ packages/material-theme/package.json
@@ -24,8 +24,7 @@
   "dependencies": {
     "@material-ui/core": "^4.12.3",
     "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
-    "@viewstools/use-masked-input": "^2.0.1"
+    "@material-ui/lab": "^4.0.0-alpha.60"
   },
   "devDependencies": {
     "react": "^16.12.0",
